Refactoring and inline documentation.
